Add files via upload
It is an original dataset of plastic input via rivers from Lebreton et al., (2017), which is a highly-cited and model-based valuable publication about the emission of plastics from land. In this paper, only rivers around the world (over 40K ) were considered whereas it neglected the emission from beaches and marine/aquaculture activities. Thus, the model result based on the data might be lower than the actual concentration. Anyway, the expandability, flexibility and other advantages of this data really outperform its weakness.